requirements.txt
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.6.12
sentence-transformers>=2.2.2
scikit-learn>=1.0.2
scikit-image>=0.19.0
opencv-python>=4.7.0
colormath>=3.0.0
numpy>=1.23.0
Pillow>=9.5.0
faiss-cpu>=1.7.4
flask>=2.3.0
flask-cors>=4.0.0
elasticsearch>=8.8.0
python-dotenv>=1.0.0
werkzeug>=2.3.0
sentencepiece>=0.1.99 